test(kt-devnet): refactor tests to require backend suffix
Tests always end with _Memstore or _Holesky now
And justfile go test commands pattern match to run the right tests
